refactored extractors to api directory. commented out show captions in main()
